Move placeholder to edit subfolder
